+++ R/queries.R
@@ -361,6 +361,45 @@ is_cpdag <- function(cg) {
   is_it
 }
 
+#' @title Is the `caugi` graph an MPDAG?
+#'
+#' @description Checks if the given `caugi` graph is a
+#' Meek-closed Partially Directed Acyclic Graph (MPDAG), i.e. a PDAG where no
+#' additional edge orientations are implied by Meek's rules (R1--R4).
+#'
+#' @details
+#' If the graph is not PDAG-compatible, the function returns `FALSE`.
+#'
+#' @param cg A `caugi` object.
+#'
+#' @returns A logical value indicating whether the graph is an MPDAG.
+#'
+#' @examples
+#' cg_mpdag <- caugi(
+#'   A %-->% B,
+#'   C %-->% B,
+#'   class = "PDAG"
+#' )
+#' is_mpdag(cg_mpdag) # TRUE
+#'
+#' cg_not_mpdag <- caugi(
+#'   A %---% B,
+#'   A %-->% C,
+#'   C %-->% B,
+#'   class = "PDAG"
+#' )
+#' is_mpdag(cg_not_mpdag) # FALSE
+#'
+#' @family queries
+#' @concept queries
+#'
+#' @export
+is_mpdag <- function(cg) {
+  is_caugi(cg, throw_error = TRUE)
+
+  rs_is_mpdag(cg@session)
+}
+
 #' @title Is the `caugi` graph an UG?
 #'
 #' @description Checks if the given `caugi` graph is an undirected graph (UG).